product_service/requirements.txt
grpcio-tools==1.21.1
mysqlclient==1.4.2.post1
SQLAlchemy==1.3.4